Successfully migrated over all functions from manager to index.js and got .then statements to work correctly
