+++ src/page-template.js
@@ -3,8 +3,32 @@
 
 let employeeArr = [];
 
-const genCard = (employeeCard) => {
-    employeeArr.push(employeeCard);
+const formatManager = (element) => {
+    let managerEl = element;
+    const { name, id, email, officeNum} = managerEl;
+    let role = managerEl.getRole();
+    return `<div class="col pb-3 pr-3 mb-3 bg-body rounded"><div class="card shadow" style="width: 18rem;"><div class="card-body bg-primary"><h5 class="card-title text-light">${name}/h5><p class="card-text text-light">${role}</p></div><div class=" bg-lgray py-3 px-1" style="width: 18rem;"><ul class="list-group list-group-flush border bd-list mx-auto" style="width: 15rem;"><li class="list-group-item">ID: ${id}</li><li class="list-group-item">Email: <a href="mailto:${email}">${email}</a></li><li class="list-group-item">Office Number: ${officeNum}</li></ul></div></div></div>`
+//     return `<div class="col pb-3 pr-3 mb-3 bg-body rounded">
+//     <div class="card shadow" style="width: 18rem;">
+//         <div class="card-body bg-primary">
+//           <h5 class="card-title text-light">${name}/h5>
+//           <p class="card-text text-light">${role}</p>
+//         </div>
+//         <div class=" bg-lgray py-3 px-1" style="width: 18rem;">
+//             <ul class="list-group list-group-flush border bd-list mx-auto" style="width: 15rem;">
+//                 <li class="list-group-item">ID: ${id}</li>
+//                 <li class="list-group-item">Email: <a href="mailto:${email}">${email}</a></li>
+//                 <li class="list-group-item">Office Number: ${officeNum}</li>
+//             </ul>
+//         </div>
+//       </div>
+// </div>`
+}
+
+// Will need switch or if statements to check for employee type
+const genCard = (employee) => {
+    const empHTML = formatManager(employee)
+    employeeArr.push(empHTML);
     console.log("Employee Array", employeeArr);
 }
